+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Change Log
+
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 This change log adheres to standards from [Keep a CHANGELOG](https://keepachangelog.com).
@@ -21,9 +22,11 @@ This change log adheres to standards from [Keep a CHANGELOG](https://keepachange
 * [`no-invalid-html-attribute`]: substitute placeholders in suggestion messages ([#3759][] @mdjermanovic)
 * [`sort-prop-types`]: single line type ending without semicolon ([#3784][] @akulsr0)
 * [`require-default-props`]: report when required props have default value ([#3785][] @akulsr0)
+* [`function-component-definition`], [`boolean-prop-naming`], [`jsx-first-prop-new-line`], [`jsx-props-no-multi-spaces`], `propTypes`: use type args ([#3629][] @HenryBrown0)
 
 ### Changed
 * [Refactor] `variableUtil`: Avoid creating a single flat variable scope for each lookup ([#3782][] @DanielRosenwasser)
+* [Tests] add @typescript-eslint/parser v6 ([#3629][] @HenryBrown0)
 
 [7.35.0]: https://github.com/jsx-eslint/eslint-plugin-react/compare/v7.34.4...v7.35.0
 [#3785]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3785
@@ -35,8 +38,10 @@ This change log adheres to standards from [Keep a CHANGELOG](https://keepachange
 [#3771]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3771
 [#3759]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3759
 [#3748]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3748
+[#3746]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3746
 [#3724]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3724
 [#3694]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3694
+[#3629]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3629
 
 ## [7.34.4] - 2024.07.13